全ファイルの作成と、i_oの移動完了
